Removed dependency on WinRar.
Instead of WinRar we're now using the 7z self extracting mechanism. The header.sfx is a customized version of the 7zS.sfx file from the 9.20 distribution. The customization is that it contains a custom icon and a manifest so that Windows compatibility doesn't try to run the resulting executable as administrator.